# Source model independent bins