--- a/debian/changelog
+++ b/debian/changelog
@@ -1,4 +1,6 @@
   [Michele Simionato]
+  * Used uniform bins for disaggregation (before they were potentially
+    different across realizations)
   * Improved the error message if the user forgets both sites and sites.csv
     in a calculation starting from predetermined GMFs
   * Improved the error message if the user specifies a non-existing file in