--- a/openquake/calculators/disaggregation.py
+++ b/openquake/calculators/disaggregation.py
@@ -178,14 +178,37 @@ producing too small PoEs.'''
         src_filter = SourceFilter(sitecol, oq.maximum_distance)
         R = len(self.rlzs_assoc.realizations)
         max_poe = numpy.zeros(R, oq.imt_dt())
+
+        # build trt_edges
+        trts = tuple(sorted(set(sg.trt for smodel in self.csm.source_models
+                                for sg in smodel.src_groups)))
+
+        # build mag_edges
+        min_mag = min(sg.min_mag for smodel in self.csm.source_models
+                      for sg in smodel.src_groups)
+        max_mag = max(sg.max_mag for smodel in self.csm.source_models
+                      for sg in smodel.src_groups)
+        mag_edges = mag_bin_width * numpy.arange(
+            int(numpy.floor(min_mag / mag_bin_width)),
+            int(numpy.ceil(max_mag / mag_bin_width) + 1))
+
+        # build dist_edges, lon_edges, lat_edges per sid
+        for sid in self.sitecol.sids:
+            bb = bb_dict[sid]
+            if not bb:
+                logging.info('site %d was too far, skipping disaggregation',
+                             sid)
+                continue
+            dist_edges, lon_edges, lat_edges = bb.bins_edges(
+                oq.distance_bin_width, oq.coordinate_bin_width)
+            self.bin_edges[sid] = bs = (
+                mag_edges, dist_edges, lon_edges, lat_edges, eps_edges)
+            shape = disagg.BinData(
+                *[len(edges) - 1 for edges in bs] + [len(trts)])
+            logging.info('%s for sid %d', shape, sid)
+
         for smodel in self.csm.source_models:
             sm_id = smodel.ordinal
-            trt_names = tuple(sorted(mod.trt for mod in smodel.src_groups))
-            max_mag = max(mod.max_mag for mod in smodel.src_groups)
-            min_mag = min(mod.min_mag for mod in smodel.src_groups)
-            mag_edges = mag_bin_width * numpy.arange(
-                int(numpy.floor(min_mag / mag_bin_width)),
-                int(numpy.ceil(max_mag / mag_bin_width) + 1))
             for i, site in enumerate(sitecol):
                 sid = sitecol.sids[i]
                 curve = curves[i]
@@ -203,14 +226,6 @@ producing too small PoEs.'''
                         site.location)
                     continue
 
-                dist_edges, lon_edges, lat_edges = bb.bins_edges(
-                    oq.distance_bin_width, oq.coordinate_bin_width)
-                self.bin_edges[sm_id, sid] = bs = (
-                    mag_edges, dist_edges, lon_edges, lat_edges, eps_edges)
-                shape = disagg.BinData(
-                    *[len(edges) - 1 for edges in bs] + [len(trt_names)])
-                logging.info('%s for model %d, sid %d', shape, sm_id, sid)
-
             # check for too big poes_disagg
             for poe in oq.poes_disagg:
                 for rlz in self.rlzs_assoc.rlzs_by_smodel[sm_id]:
@@ -221,9 +236,6 @@ producing too small PoEs.'''
                             raise ValueError(self.POE_TOO_BIG % (
                                 poe, sm_id, smodel.name, min_poe, rlzi, imt))
 
-            bin_edges = {sid: self.bin_edges[sm_id, sid]
-                         for sid in sitecol.sids
-                         if (sm_id, sid) in self.bin_edges}
             for sg in smodel.src_groups:
                 split_sources = []
                 for src in sg:
@@ -241,8 +253,8 @@ producing too small PoEs.'''
                     curve) for curve in curves]
                 for srcs in split_in_blocks(split_sources, nblocks):
                     all_args.append(
-                        (src_filter, srcs, cmaker, imls, trt_names,
-                         bin_edges, oq, mon))
+                        (src_filter, srcs, cmaker, imls, trts,
+                         self.bin_edges, oq, mon))
 
         results = parallel.Starmap(compute_disagg, all_args).reduce(
             self.agg_result)
@@ -266,7 +278,7 @@ producing too small PoEs.'''
         # is saved this method can be run sequentially on the controller node
         for key, probs in sorted(results.items()):
             sid, rlz_id, poe, imt, iml, trt_names = key
-            edges = self.bin_edges[sm_id[rlz_id], sid]
+            edges = self.bin_edges[sid]
             self.save_disagg_result(
                 sid, edges, trt_names, probs, rlz_id,
                 self.oqparam.investigation_time, imt, iml, poe)