--- a/openquake/qa_tests_data/disagg/case_2/expected_output/rlz-2-PGA-0.0-0.0.xml
+++ b/openquake/qa_tests_data/disagg/case_2/expected_output/rlz-2-PGA-0.0-0.0.xml
@@ -15,7 +15,7 @@ xmlns:gml="http://www.opengis.net/gml"
     lonBinEdges="-5.000000E-01, -4.000000E-01, -3.000000E-01, -2.000000E-01, -1.000000E-01, 0.000000E+00, 1.000000E-01, 2.000000E-01, 3.000000E-01, 4.000000E-01, 5.000000E-01"
     magBinEdges="5.000000E+00, 5.100000E+00, 5.200000E+00, 5.300000E+00, 5.400000E+00, 5.500000E+00, 5.600000E+00, 5.700000E+00, 5.800000E+00, 5.900000E+00, 6.000000E+00, 6.100000E+00, 6.200000E+00, 6.300000E+00, 6.400000E+00, 6.500000E+00"
     sourceModelTreePath="source_model_2"
-    tectonicRegionTypes="Active Shallow Crust"
+    tectonicRegionTypes="Active Shallow Crust, Subduction IntraSlab"
     >
         <disaggMatrix
         dims="15"
@@ -55,12 +55,13 @@ xmlns:gml="http://www.opengis.net/gml"
             <prob index="7" value="2.698025E-08"/>
         </disaggMatrix>
         <disaggMatrix
-        dims="1"
+        dims="2"
         iml="1.000000E-01"
         poE="1.836464E-02"
         type="TRT"
         >
             <prob index="0" value="1.836464E-02"/>
+            <prob index="1" value="0.000000E+00"/>
         </disaggMatrix>
         <disaggMatrix
         dims="15,8"
@@ -2531,111 +2532,211 @@ xmlns:gml="http://www.opengis.net/gml"
             <prob index="14,9,9" value="0.000000E+00"/>
         </disaggMatrix>
         <disaggMatrix
-        dims="10,10,1"
+        dims="10,10,2"
         iml="1.000000E-01"
         poE="1.836464E-02"
         type="Lon,Lat,TRT"
         >
             <prob index="0,0,0" value="1.609876E-06"/>
+            <prob index="0,0,1" value="0.000000E+00"/>
             <prob index="0,1,0" value="0.000000E+00"/>
+            <prob index="0,1,1" value="0.000000E+00"/>
             <prob index="0,2,0" value="0.000000E+00"/>
+            <prob index="0,2,1" value="0.000000E+00"/>
             <prob index="0,3,0" value="0.000000E+00"/>
+            <prob index="0,3,1" value="0.000000E+00"/>
             <prob index="0,4,0" value="0.000000E+00"/>
+            <prob index="0,4,1" value="0.000000E+00"/>
             <prob index="0,5,0" value="0.000000E+00"/>
+            <prob index="0,5,1" value="0.000000E+00"/>
             <prob index="0,6,0" value="0.000000E+00"/>
+            <prob index="0,6,1" value="0.000000E+00"/>
             <prob index="0,7,0" value="0.000000E+00"/>
+            <prob index="0,7,1" value="0.000000E+00"/>
             <prob index="0,8,0" value="0.000000E+00"/>
+            <prob index="0,8,1" value="0.000000E+00"/>
             <prob index="0,9,0" value="0.000000E+00"/>
+            <prob index="0,9,1" value="0.000000E+00"/>
             <prob index="1,0,0" value="0.000000E+00"/>
+            <prob index="1,0,1" value="0.000000E+00"/>
             <prob index="1,1,0" value="7.011501E-05"/>
+            <prob index="1,1,1" value="0.000000E+00"/>
             <prob index="1,2,0" value="0.000000E+00"/>
+            <prob index="1,2,1" value="0.000000E+00"/>
             <prob index="1,3,0" value="0.000000E+00"/>
+            <prob index="1,3,1" value="0.000000E+00"/>
             <prob index="1,4,0" value="0.000000E+00"/>
+            <prob index="1,4,1" value="0.000000E+00"/>
             <prob index="1,5,0" value="0.000000E+00"/>
+            <prob index="1,5,1" value="0.000000E+00"/>
             <prob index="1,6,0" value="0.000000E+00"/>
+            <prob index="1,6,1" value="0.000000E+00"/>
             <prob index="1,7,0" value="0.000000E+00"/>
+            <prob index="1,7,1" value="0.000000E+00"/>
             <prob index="1,8,0" value="0.000000E+00"/>
+            <prob index="1,8,1" value="0.000000E+00"/>
             <prob index="1,9,0" value="0.000000E+00"/>
+            <prob index="1,9,1" value="0.000000E+00"/>
             <prob index="2,0,0" value="0.000000E+00"/>
+            <prob index="2,0,1" value="0.000000E+00"/>
             <prob index="2,1,0" value="0.000000E+00"/>
+            <prob index="2,1,1" value="0.000000E+00"/>
             <prob index="2,2,0" value="2.832871E-04"/>
+            <prob index="2,2,1" value="0.000000E+00"/>
             <prob index="2,3,0" value="0.000000E+00"/>
+            <prob index="2,3,1" value="0.000000E+00"/>
             <prob index="2,4,0" value="0.000000E+00"/>
+            <prob index="2,4,1" value="0.000000E+00"/>
             <prob index="2,5,0" value="0.000000E+00"/>
+            <prob index="2,5,1" value="0.000000E+00"/>
             <prob index="2,6,0" value="0.000000E+00"/>
+            <prob index="2,6,1" value="0.000000E+00"/>
             <prob index="2,7,0" value="0.000000E+00"/>
+            <prob index="2,7,1" value="0.000000E+00"/>
             <prob index="2,8,0" value="0.000000E+00"/>
+            <prob index="2,8,1" value="0.000000E+00"/>
             <prob index="2,9,0" value="0.000000E+00"/>
+            <prob index="2,9,1" value="0.000000E+00"/>
             <prob index="3,0,0" value="0.000000E+00"/>
+            <prob index="3,0,1" value="0.000000E+00"/>
             <prob index="3,1,0" value="0.000000E+00"/>
+            <prob index="3,1,1" value="0.000000E+00"/>
             <prob index="3,2,0" value="0.000000E+00"/>
+            <prob index="3,2,1" value="0.000000E+00"/>
             <prob index="3,3,0" value="1.085338E-03"/>
+            <prob index="3,3,1" value="0.000000E+00"/>
             <prob index="3,4,0" value="0.000000E+00"/>
+            <prob index="3,4,1" value="0.000000E+00"/>
             <prob index="3,5,0" value="0.000000E+00"/>
+            <prob index="3,5,1" value="0.000000E+00"/>
             <prob index="3,6,0" value="0.000000E+00"/>
+            <prob index="3,6,1" value="0.000000E+00"/>
             <prob index="3,7,0" value="0.000000E+00"/>
+            <prob index="3,7,1" value="0.000000E+00"/>
             <prob index="3,8,0" value="0.000000E+00"/>
+            <prob index="3,8,1" value="0.000000E+00"/>
             <prob index="3,9,0" value="0.000000E+00"/>
+            <prob index="3,9,1" value="0.000000E+00"/>
             <prob index="4,0,0" value="0.000000E+00"/>
+            <prob index="4,0,1" value="0.000000E+00"/>
             <prob index="4,1,0" value="0.000000E+00"/>
+            <prob index="4,1,1" value="0.000000E+00"/>
             <prob index="4,2,0" value="0.000000E+00"/>
+            <prob index="4,2,1" value="0.000000E+00"/>
             <prob index="4,3,0" value="0.000000E+00"/>
+            <prob index="4,3,1" value="0.000000E+00"/>
             <prob index="4,4,0" value="5.053343E-03"/>
+            <prob index="4,4,1" value="0.000000E+00"/>
             <prob index="4,5,0" value="0.000000E+00"/>
+            <prob index="4,5,1" value="0.000000E+00"/>
             <prob index="4,6,0" value="0.000000E+00"/>
+            <prob index="4,6,1" value="0.000000E+00"/>
             <prob index="4,7,0" value="0.000000E+00"/>
+            <prob index="4,7,1" value="0.000000E+00"/>
             <prob index="4,8,0" value="0.000000E+00"/>
+            <prob index="4,8,1" value="0.000000E+00"/>
             <prob index="4,9,0" value="0.000000E+00"/>
+            <prob index="4,9,1" value="0.000000E+00"/>
             <prob index="5,0,0" value="0.000000E+00"/>
+            <prob index="5,0,1" value="0.000000E+00"/>
             <prob index="5,1,0" value="0.000000E+00"/>
+            <prob index="5,1,1" value="0.000000E+00"/>
             <prob index="5,2,0" value="0.000000E+00"/>
+            <prob index="5,2,1" value="0.000000E+00"/>
             <prob index="5,3,0" value="0.000000E+00"/>
+            <prob index="5,3,1" value="0.000000E+00"/>
             <prob index="5,4,0" value="0.000000E+00"/>
+            <prob index="5,4,1" value="0.000000E+00"/>
             <prob index="5,5,0" value="1.036744E-02"/>
+            <prob index="5,5,1" value="0.000000E+00"/>
             <prob index="5,6,0" value="0.000000E+00"/>
+            <prob index="5,6,1" value="0.000000E+00"/>
             <prob index="5,7,0" value="0.000000E+00"/>
+            <prob index="5,7,1" value="0.000000E+00"/>
             <prob index="5,8,0" value="0.000000E+00"/>
+            <prob index="5,8,1" value="0.000000E+00"/>
             <prob index="5,9,0" value="0.000000E+00"/>
+            <prob index="5,9,1" value="0.000000E+00"/>
             <prob index="6,0,0" value="0.000000E+00"/>
+            <prob index="6,0,1" value="0.000000E+00"/>
             <prob index="6,1,0" value="0.000000E+00"/>
+            <prob index="6,1,1" value="0.000000E+00"/>
             <prob index="6,2,0" value="0.000000E+00"/>
+            <prob index="6,2,1" value="0.000000E+00"/>
             <prob index="6,3,0" value="0.000000E+00"/>
+            <prob index="6,3,1" value="0.000000E+00"/>
             <prob index="6,4,0" value="0.000000E+00"/>
+            <prob index="6,4,1" value="0.000000E+00"/>
             <prob index="6,5,0" value="0.000000E+00"/>
+            <prob index="6,5,1" value="0.000000E+00"/>
             <prob index="6,6,0" value="1.212070E-03"/>
+            <prob index="6,6,1" value="0.000000E+00"/>
             <prob index="6,7,0" value="0.000000E+00"/>
+            <prob index="6,7,1" value="0.000000E+00"/>
             <prob index="6,8,0" value="0.000000E+00"/>
+            <prob index="6,8,1" value="0.000000E+00"/>
             <prob index="6,9,0" value="0.000000E+00"/>
+            <prob index="6,9,1" value="0.000000E+00"/>
             <prob index="7,0,0" value="0.000000E+00"/>
+            <prob index="7,0,1" value="0.000000E+00"/>
             <prob index="7,1,0" value="0.000000E+00"/>
+            <prob index="7,1,1" value="0.000000E+00"/>
             <prob index="7,2,0" value="0.000000E+00"/>
+            <prob index="7,2,1" value="0.000000E+00"/>
             <prob index="7,3,0" value="0.000000E+00"/>
+            <prob index="7,3,1" value="0.000000E+00"/>
             <prob index="7,4,0" value="0.000000E+00"/>
+            <prob index="7,4,1" value="0.000000E+00"/>
             <prob index="7,5,0" value="0.000000E+00"/>
+            <prob index="7,5,1" value="0.000000E+00"/>
             <prob index="7,6,0" value="0.000000E+00"/>
+            <prob index="7,6,1" value="0.000000E+00"/>
             <prob index="7,7,0" value="3.139759E-04"/>
+            <prob index="7,7,1" value="0.000000E+00"/>
             <prob index="7,8,0" value="0.000000E+00"/>
+            <prob index="7,8,1" value="0.000000E+00"/>
             <prob index="7,9,0" value="0.000000E+00"/>
+            <prob index="7,9,1" value="0.000000E+00"/>
             <prob index="8,0,0" value="0.000000E+00"/>
+            <prob index="8,0,1" value="0.000000E+00"/>
             <prob index="8,1,0" value="0.000000E+00"/>
+            <prob index="8,1,1" value="0.000000E+00"/>
             <prob index="8,2,0" value="0.000000E+00"/>
+            <prob index="8,2,1" value="0.000000E+00"/>
             <prob index="8,3,0" value="0.000000E+00"/>
+            <prob index="8,3,1" value="0.000000E+00"/>
             <prob index="8,4,0" value="0.000000E+00"/>
+            <prob index="8,4,1" value="0.000000E+00"/>
             <prob index="8,5,0" value="0.000000E+00"/>
+            <prob index="8,5,1" value="0.000000E+00"/>
             <prob index="8,6,0" value="0.000000E+00"/>
+            <prob index="8,6,1" value="0.000000E+00"/>
             <prob index="8,7,0" value="0.000000E+00"/>
+            <prob index="8,7,1" value="0.000000E+00"/>
             <prob index="8,8,0" value="7.789822E-05"/>
+            <prob index="8,8,1" value="0.000000E+00"/>
             <prob index="8,9,0" value="0.000000E+00"/>
+            <prob index="8,9,1" value="0.000000E+00"/>
             <prob index="9,0,0" value="0.000000E+00"/>
+            <prob index="9,0,1" value="0.000000E+00"/>
             <prob index="9,1,0" value="0.000000E+00"/>
+            <prob index="9,1,1" value="0.000000E+00"/>
             <prob index="9,2,0" value="0.000000E+00"/>
+            <prob index="9,2,1" value="0.000000E+00"/>
             <prob index="9,3,0" value="0.000000E+00"/>
+            <prob index="9,3,1" value="0.000000E+00"/>
             <prob index="9,4,0" value="0.000000E+00"/>
+            <prob index="9,4,1" value="0.000000E+00"/>
             <prob index="9,5,0" value="0.000000E+00"/>
+            <prob index="9,5,1" value="0.000000E+00"/>
             <prob index="9,6,0" value="0.000000E+00"/>
+            <prob index="9,6,1" value="0.000000E+00"/>
             <prob index="9,7,0" value="0.000000E+00"/>
+            <prob index="9,7,1" value="0.000000E+00"/>
             <prob index="9,8,0" value="0.000000E+00"/>
+            <prob index="9,8,1" value="0.000000E+00"/>
             <prob index="9,9,0" value="1.941774E-06"/>
+            <prob index="9,9,1" value="0.000000E+00"/>
         </disaggMatrix>
     </disaggMatrices>
 </nrml>