Closes https://github.com/gem/oq-engine/issues/3236. The changes in the test case_2 are easily explained. It is a test with two source models, source_model_1.xml (with tectonic region types "Active Shallow Crust" and "Subduction IntraSlab") and source_model_2.xml (with a single TRT "Active Shallow Crust"). In master there are distinct bins for source_model_1.xml and source_model_2.xml; in particular the bins for the second source model are missing the TRT  "Subduction IntraSlab". In this branch instead the same bins, including "Subduction IntraSlab" are used everywhere. This explains the change in the exported files for the realizations 2 and 3 that corresponds to source_model_2.xml (the first source model instead corresponds to the realizations 0 and 1 which are unchanged). The magnitude, distance and lon, lat bins are already uniform across source models in this test, this is why there are no other differences. In *case_master* the magnitude numbers are different because the magnitude bins are different. From the log one gets:

before:
BinData(mags=5, dists=4, lons=2, lats=1, eps=4, trts=2) for model 0, sid 0
BinData(mags=5, dists=3, lons=2, lats=1, eps=4, trts=2) for model 0, sid 1
BinData(mags=4, dists=4, lons=2, lats=1, eps=4, trts=2) for model 1, sid 0
BinData(mags=4, dists=3, lons=2, lats=1, eps=4, trts=2) for model 1, sid 1

after:
BinData(mags=6, dists=4, lons=2, lats=1, eps=4, trts=2) for sid 0
BinData(mags=6, dists=3, lons=2, lats=1, eps=4, trts=2) for sid 1

For some reason, this change doubled the speed of the Armenia calculation:
```
   rgee@wilson:~$ oq show performance 16454  -- master
   ================================== ========= ========= ==========
   operation                          time_sec  memory_mb counts    
   ================================== ========= ========= ==========
   total compute_disagg               951,034   43        2,187     
   arranging bins                     775,391   62        13,920    
   collecting bins                    175,284   61        2,187     
   disaggregate_pne                   104,438   0.0       95,850,776
   total pmap_from_grp                88,787    7.379     2,307     
   making contexts                    25,930    0.0       25,163,900
   total runtime                      3,053     1,551     1      

   rgee@wilson:~$ oq show performance 16494  -- this branch
   operation                          time_sec  memory_mb counts
   ================================== ========= ========= ==========
   total compute_disagg               492,507   34        2,187
   arranging bins                     320,498   24        13,920
   collecting bins                    171,994   40        2,187
   disaggregate_pne                   103,243   0.0       95,850,776
   total pmap_from_grp                88,373    7.391     2,307
   making contexts                    25,766    0.0       25,163,900
   total runtime                      1,968     1,533     1
```