@micheles 2017-11-15
Unfortunately this cannot be case, because this is the loop sending the arguments to the tasks, while the improvement is in the operations *inside* the tasks.